+++ packages/admin/admin-babel-preset/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/admin-babel-preset
 
+## 6.4.0
+
 ## 6.3.0
 
 ## 6.2.1